backend/services/osint_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]
redis==5.0.1
aiohttp==3.9.1
asyncio==3.4.3
beautifulsoup4==4.12.2
lxml==5.1.0
selenium==4.17.2
playwright==1.41.1
Pillow==10.2.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
exifread==3.0.0
phonenumbers==8.13.28
email-validator==2.1.0
dnspython==2.4.2
python-whois==0.8.0
tweepy==4.14.0
discord.py==2.3.2
instaloader==4.10.3
linkedin-api==2.0.0a5
telethon==1.34.0
httpx==0.26.0
numpy==1.26.3
scikit-learn==1.4.0
pandas==2.1.4
cryptography==42.0.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
fake-useragent==1.4.0
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
tqdm==4.66.1
colorama==0.4.6
python-dotenv==1.0.0
